fix: compile against RuneLite 1.13 long GE price API and widen GP money to long
RuneLite 1.13 (api: update 241) widens GrandExchangeOffer.getPrice()/getSpent()
and ItemManager.getItemPrice() to long ahead of Jagex's Sep 30 max-cash update,
after which a single GE price can exceed Integer.MAX_VALUE. master failed to
compile against 1.13.0-SNAPSHOT with 10 lossy long->int conversions.

Every receiving local, field and parameter is widened to long instead of cast,
so the plugin compiles against both latest.release (1.12, int widens
implicitly) and latest.integration (1.13) and can target plugin-hub master.
Nothing narrows a GE price or spent back to int.

The same sweep widens every GP value that would break above 2.147b (#1364):
- Gson response and request DTOs (FlipRecommendation, ActiveFlip, FlipAnalysis,
  FavoriteItem, OfferAdvice*, FlipAdjustment*, SellPriceCheck*, Transaction/
  HistoryBackfill/BankItem requests, FlipFinderResponse.cashStack). An int field
  threw JsonSyntaxException and dropped the whole response.
- WikiPrice parsing uses getAsLong.
- Silent overflow/saturation: slot tooltip P&L, Flip Assist total profit,
  FocusedFlip adjusted profit, GeTax breakeven, SmartSellPricer min profitable
  (+1 wrapped negative), Flip Finder 5% markup, SessionStats gp/hr.
- GeTax.breakevenSellPrice starts from min(closed form, buy + cap), so a capped
  price no longer walks down millions of steps.
- GpUtils.parseDigits and GEHistoryService parsing return long; GE History rows
  above int32 are no longer dropped.
- Formatter/util signatures, session/adjustment/stale-offer price maps.
- Persisted state: OfferRecord.price, AutoRecommendService buyPrices,
  ExitTrades buyBasis. Old int JSON still loads into the long fields; a
  downgrade after writing >int32 values loses that state (restores already
  catch the exception).

Left int by design: quantities, item ids, slots, GE tax (capped at 5m), the
coin-stack cash stack and inventory GP (cash stack cap unchanged), config
thresholds, and the GE_NEWOFFER_PRICE varbit read (a varbit cannot carry more
than 32 bits; negative reads already clamp to 0).

Tests stub RuneLite money getters through GpAnswer so the suite compiles and
passes on both 1.12 and 1.13.

Refs Flip-Smart/flip-smart#1363, Flip-Smart/flip-smart#1364

+++ src/main/java/com/flipsmart/ActiveOfferAdvisorService.java
@@ -21,11 +21,11 @@ public class ActiveOfferAdvisorService
 	@Getter
 	public static final class CourierState
 	{
-		private final Integer previousPositionMargin;
+		private final Long previousPositionMargin;
 		private final int consecutiveMarginDecreases;
 		private final double cumulativeMarginReductionPct;
 
-		public CourierState(Integer previousPositionMargin, int consecutiveMarginDecreases, double cumulativeMarginReductionPct)
+		public CourierState(Long previousPositionMargin, int consecutiveMarginDecreases, double cumulativeMarginReductionPct)
 		{
 			this.previousPositionMargin = previousPositionMargin;
 			this.consecutiveMarginDecreases = consecutiveMarginDecreases;
@@ -127,7 +127,7 @@ public class ActiveOfferAdvisorService
 	static OfferAdviceRequest buildSnapshot(
 		OfferRecord offer,
 		WikiPrice market,
-		Integer userAvgBuyPrice,
+		Long userAvgBuyPrice,
 		Integer dailyVolume)
 	{
 		return buildSnapshot(offer, market, userAvgBuyPrice, dailyVolume, null, CourierState.EMPTY);
@@ -139,7 +139,7 @@ public class ActiveOfferAdvisorService
 	 * off we relay null / EMPTY so the backend runs only the base wait / move-down / exit advice.
 	 * Static + package-private so the gate stays unit-testable.
 	 */
-	static Integer relayedMargin(boolean aggressive, Integer originalMargin)
+	static Long relayedMargin(boolean aggressive, Long originalMargin)
 	{
 		return aggressive ? originalMargin : null;
 	}
@@ -152,9 +152,9 @@ public class ActiveOfferAdvisorService
 	static OfferAdviceRequest buildSnapshot(
 		OfferRecord offer,
 		WikiPrice market,
-		Integer userAvgBuyPrice,
+		Long userAvgBuyPrice,
 		Integer dailyVolume,
-		Integer originalMargin,
+		Long originalMargin,
 		CourierState courier)
 	{
 		boolean isSell = !offer.isBuy();
